Move period status helpers to period-status module
