+++ src/application/summary-logs/validate.js
@@ -16,7 +16,6 @@ import {
 import { PermanentError } from '#server/queue-consumer/permanent-error.js'
 
 import { transformFromSummaryLog } from '#application/waste-records/transform-from-summary-log.js'
-import { VERSION_STATUS } from '#domain/waste-records/model.js'
 import { SUMMARY_LOG_META_FIELDS } from '#domain/summary-logs/meta-fields.js'
 import {
   PROCESSING_TYPE_TABLES,
@@ -30,7 +29,7 @@ import {
   countByWasteRecordType,
   mergeLoads
 } from './load-counts.js'
-import { classifyByPeriodStatus } from './period-status.js'
+import { computeLoadsByPeriodStatus } from './period-status.js'
 import {
   logValidationIssues,
   MAX_VALIDATION_ISSUES
@@ -603,141 +602,6 @@ const classifyLoads = ({
   })
 
   return { loads, loadsByWasteRecordType }
-}
-
-/**
- * Computes the transaction amount for a record's data via classifyForWasteBalance.
- * Returns 0 if the record is not INCLUDED.
- *
- * @param {import('#domain/summary-logs/table-schemas/index.js').TableSchema | null} schema
- * @param {Record<string, any>} data
- * @param {Registration} registration
- * @returns {number}
- */
-const getTransactionAmount = (schema, data, registration) => {
-  const result = schema?.classifyForWasteBalance?.(data, {
-    accreditation: registration.accreditation,
-    overseasSites: ORS_VALIDATION_DISABLED
-  })
-  return result?.outcome === ROW_OUTCOME.INCLUDED ? result.transactionAmount : 0
-}
-
-/**
- * Builds a map of rowId => waste balance impact for included waste-balance records.
- *
- * For added records, the impact is the full transaction amount.
- * For adjusted records, the impact is the delta: new amount minus old amount.
- * This reflects the actual change each record makes to the waste balance.
- *
- * @param {ValidatedWasteRecord[]} wasteBalanceRecords
- * @param {Registration} registration
- * @param {string} processingType
- * @param {string} summaryLogId
- * @param {Map<string, import('#domain/waste-records/model.js').WasteRecord>} existingRecordsMap - keyed by "type:rowId"
- * @returns {Map<string, number>}
- */
-export const buildTransactionAmounts = (
-  wasteBalanceRecords,
-  registration,
-  processingType,
-  summaryLogId,
-  existingRecordsMap
-) => {
-  const amounts = new Map()
-  for (const { record, outcome } of wasteBalanceRecords) {
-    if (outcome !== ROW_OUTCOME.INCLUDED) {
-      continue
-    }
-    const schema = findSchemaForProcessingType(processingType, record.type)
-    const newAmount = getTransactionAmount(schema, record.data, registration)
-    if (newAmount === 0) {
-      continue
-    }
-
-    const key = `${record.type}:${record.rowId}`
-    const lastVersion = record.versions[record.versions.length - 1]
-    const isAdjusted =
-      lastVersion.summaryLog?.id === summaryLogId &&
-      lastVersion.status === VERSION_STATUS.UPDATED
-
-    if (isAdjusted) {
-      const existing = existingRecordsMap.get(key)
-      const oldAmount = existing
-        ? getTransactionAmount(schema, existing.data, registration)
-        : 0
-      amounts.set(key, newAmount - oldAmount)
-    } else {
-      amounts.set(key, newAmount)
-    }
-  }
-  return amounts
-}
-
-/**
- * Computes loadsByPeriodStatus for a validated summary log.
- * Returns null if the summary log is not validated or if the reports lookup fails.
- *
- * @param {Object} params
- * @param {ValidatedWasteRecord[]} params.wasteRecords
- * @param {ValidatedWasteRecord[]} params.wasteBalanceRecords
- * @param {string} params.summaryLogId
- * @param {Registration} params.registration
- * @param {string} params.processingType
- * @param {Map<string, import('#domain/waste-records/model.js').WasteRecord>} params.existingRecordsMap
- * @param {import('#reports/repository/port.js').ReportsRepository} params.reportsRepository
- * @param {string} params.organisationId
- * @param {string} params.registrationId
- * @param {string} params.loggingContext
- * @param {TypedLogger} params.logger
- * @returns {Promise<import('./period-status.js').LoadsByPeriodStatus | null>}
- */
-const computeLoadsByPeriodStatus = async ({
-  wasteRecords,
-  wasteBalanceRecords,
-  summaryLogId,
-  registration,
-  processingType,
-  existingRecordsMap,
-  reportsRepository,
-  organisationId,
-  registrationId,
-  loggingContext,
-  logger
-}) => {
-  try {
-    const submittedReports = await reportsRepository.findPeriodicReports({
-      organisationId,
-      registrationId
-    })
-
-    const transactionAmounts = buildTransactionAmounts(
-      wasteBalanceRecords,
-      registration,
-      processingType,
-      summaryLogId,
-      existingRecordsMap
-    )
-
-    return classifyByPeriodStatus({
-      wasteRecords,
-      wasteBalanceRecords,
-      summaryLogId,
-      registration,
-      submittedReports,
-      tableSchemas: PROCESSING_TYPE_TABLES[processingType],
-      transactionAmounts
-    })
-  } catch (err) {
-    logger.warn({
-      message: `Failed to classify loads by period status: ${loggingContext}`,
-      err,
-      event: {
-        category: LOGGING_EVENT_CATEGORIES.SERVER,
-        action: LOGGING_EVENT_ACTIONS.PROCESS_FAILURE
-      }
-    })
-    return null
-  }
 }
 
 /**
